STACatalogs

Documentation for STACatalogs.

STACatalogs.STACatalog

STACatalogs.bbox

STACatalogs.description

STACatalogs.eachcatalog

STACatalogs.eachitem

STACatalogs.geometry

STACatalogs.id

STACatalogs.id

STACatalogs.links

STACatalogs.properties

STACatalogs.set_cache_max_size

STACatalogs.STACatalog — Method

cat = STACatalog(url)

Open a SpatioTemporal Asset Catalog (STAC) using the provided url. The url should point to a JSON object conforming to the STAC specification.

cat behaves as a julia dictionary with all STAC subcatalogs. cat.items is a dictionary with all STAC items.

using STACatalogs
url = "https://raw.githubusercontent.com/sat-utils/sat-stac/master/test/catalog/catalog.json"
cat = STACatalog(url)
subcat = cat["stac-catalog-eo"]
subcat1 = subcat["landsat-8-l1"]
@show subcat1

item = subcat1.items["LC08_L1TP_152038_20200611_20200611_01_RT"]
@show item.assets[:B4][:href]

STACatalogs.bbox — Method

 data = bbox(item)
 Get the bounding box of STAC `item`.

STACatalogs.description — Method

 data = description(cat::Catalog)
 Get the description of STAC catalog.

STACatalogs.eachcatalog — Method

cats = eachcatalog(catalog::Catalog)

Returns resursively all subcatalogs in catalog. This can take a long time for deeply nested catalogs.

url = "https://raw.githubusercontent.com/sat-utils/sat-stac/master/test/catalog/catalog.json"

cat = STACatalog(url)
for c in eachcatalog(cat)
    @show id(c)
end

STACatalogs.eachitem — Method

cats = eachitem(catalog::Catalog)

Returns resursively all items in catalog. This can take a long time for deeply nested catalogs.

url = "https://raw.githubusercontent.com/sat-utils/sat-stac/master/test/catalog/catalog.json"

cat = STACatalog(url)
for c in eachitem(cat)
    @show id(c)
end

STACatalogs.geometry — Method

data = geometry(item) Get the geometry of STAC item as a GeoJSON object

STACatalogs.id — Method

 data = id(cat::Catalog)
 Get the identifier of STAC catalog.

STACatalogs.id — Method

 data = id(item)
 Get the identifier of STAC `item`.

STACatalogs.links — Method

 data = links(item)
 Get the links of STAC `item`.

STACatalogs.properties — Method

 data = properties(item)
 Get the properties of STAC `item`.

STACatalogs.set_cache_max_size — Method

STACatalogs.set_cache_max_size(cache_max_size::Integer)

Set the maximum number of URLs saved in cache (permanentaly). The default is 1000. The Julia session need to be restated for this change to take effect.