STAC

Documentation for STAC.

Dates.DateTime

STAC.Catalog

STAC.bbox

STAC.description

STAC.description

STAC.eachcatalog

STAC.eachitem

STAC.geometry

STAC.href

STAC.id

STAC.id

STAC.links

STAC.properties

STAC.set_cache_max_size

STAC.title

STAC.type

Dates.DateTime — Method

dt = DateTime(item)

Get the date time of STAC item as a Dates.DateTime (or nothing if this properties is not specified).

STAC.Catalog — Method

cat = STAC.Catalog(url)

Open a SpatioTemporal Asset Catalog (STAC) using the provided url. The url should point to a JSON object conforming to the STAC specification.

cat behaves as a julia dictionary with all STAC subcatalogs. cat.items is a dictionary with all STAC items.

using STAC
url = "https://raw.githubusercontent.com/sat-utils/sat-stac/master/test/catalog/catalog.json"
cat = STAC.Catalog(url)
subcat = cat["stac-catalog-eo"]
subcat1 = subcat["landsat-8-l1"]
@show subcat1

item = subcat1.items["LC08_L1TP_152038_20200611_20200611_01_RT"]
@show href(item.assets["B4"])

STAC.bbox — Method

 data = bbox(item)

Get the bounding box of STAC item.

STAC.description — Method

 data = description(asset)

Get the description of STAC asset.

STAC.description — Method

data = description(cat::Catalog)

Get the description of STAC catalog.

STAC.eachcatalog — Method

cats = eachcatalog(catalog::Catalog)

Returns resursively all subcatalogs in catalog. This can take a long time for deeply nested catalogs.

url = "https://raw.githubusercontent.com/sat-utils/sat-stac/master/test/catalog/catalog.json"

cat = STAC.Catalog(url)
for c in eachcatalog(cat)
    @show id(c)
end

STAC.eachitem — Method

cats = eachitem(catalog::Catalog)

Returns resursively all items in catalog. This can take a long time for deeply nested catalogs.

url = "https://raw.githubusercontent.com/sat-utils/sat-stac/master/test/catalog/catalog.json"

cat = STAC.Catalog(url)
for c in eachitem(cat)
    @show id(c)
end

STAC.geometry — Method

data = geometry(item)

Get the geometry of STAC item as a GeoJSON object

STAC.href — Method

 data = href(asset)

Get the URI of STAC asset.

STAC.id — Method

data = id(cat::Catalog)

Get the identifier of STAC catalog.

STAC.id — Method

 data = id(item)

Get the identifier of STAC item.

STAC.links — Method

 data = links(item)

Get the links of STAC item.

STAC.properties — Method

 data = properties(item)

Get the properties of STAC item.

STAC.set_cache_max_size — Method

STAC.set_cache_max_size(cache_max_size::Integer)

Set the maximum number of URLs saved in cache (permanentaly). The default is 1000. The Julia session need to be restated for this change to take effect.

STAC.title — Method

 data = title(asset)

Get the title of STAC asset.

STAC.type — Method

 data = type(asset)

Get the type of STAC asset.